changes on match loading
Working:
Currently, the robot is scoring the triball alliance, moving to the corner, and doing the matchload-shooting cycle, just tested 4 loads in a row.

Issues:
There are too many corrections being made, and the robot attempts to reach the point too precisely, resulting in time wastage.

Changes:

The catapult now uses a different atan function.
A flag called "setCatapultDown" has been introduced, with a default setting of true. If set to false, the catapult should get on the half position to take  triballs from the floor, holding them for scoring or releasing them at medium